Enum glean::ErrorType

#[repr(C)]
pub enum ErrorType {
    InvalidValue = 0,
    InvalidLabel = 1,
    InvalidState = 2,
    InvalidOverflow = 3,
}

The possible error types for metric recording.

Note: the cases in this enum must be kept in sync with the ones in the platform-specific code (e.g. ErrorType.kt) and with the metrics in the registry files.

Variants

InvalidValue = 0

For when the value to be recorded does not match the metric-specific restrictions

InvalidLabel = 1

For when the label of a labeled metric does not match the restrictions

InvalidState = 2

For when the metric caught an invalid state while recording

InvalidOverflow = 3

For when the value to be recorded overflows the metric-specific upper range

Implementations

impl ErrorType

pub fn as_str(&self) -> &'static str

The error type’s metric id

pub fn iter() -> impl Iterator<Item = ErrorType>

Return an iterator over all possible error types.

let errors = ErrorType::iter();
let all_errors = errors.collect::<Vec<_>>();
assert_eq!(4, all_errors.len());