Module glean::net

Handling the Glean upload logic.

This doesn't perform the actual upload but rather handles retries, upload limitations and error tracking.

Structs

HttpUploader
A simple mechanism to upload pings over HTTPS.

PingRequest
Represents a request to upload a ping.

Enums

UploadResult
The result of an attempted ping upload.

Traits

PingUploader
A description of a component used to upload pings.