Enum glean::net::UploadResult

pub enum UploadResult {
    RecoverableFailure { /* private fields */ },
    UnrecoverableFailure { /* private fields */ },
    Incapable { /* private fields */ },
    HttpStatus {
        code: i32,
    },
    Done { /* private fields */ },
}

The result of an attempted ping upload.

Variants

RecoverableFailure

A recoverable failure.

During upload something went wrong, e.g. the network connection failed. The upload should be retried at a later time.

UnrecoverableFailure

An unrecoverable upload failure.

A possible cause might be a malformed URL.

Incapable

The uploader is not capable of uploading this request due to lack of or mismatched capabilities.

e.g. The ping requires upload over OHTTP, but the uploader doesn’t support OHTTP.

HttpStatus

A HTTP response code.

This can still indicate an error, depending on the status code.

Fields

code: i32

The HTTP status code

Done

Signal that this uploader is done with work and won’t accept new work.

Implementations

impl UploadResult

pub fn get_label(&self) -> Option<&str>

Gets the label to be used in recording error counts for upload.

Returns None if the upload finished succesfully. Failures are recorded in the ping_upload_failure metric.

pub fn recoverable_failure() -> UploadResult

A recoverable failure.

During upload something went wrong, e.g. the network connection failed. The upload should be retried at a later time.

pub fn unrecoverable_failure() -> UploadResult

An unrecoverable upload failure.

A possible cause might be a malformed URL.

pub fn incapable() -> UploadResult

The uploader is not capable of uploading this request due to lack of or mismatched capabilities.

e.g. The ping requires upload over OHTTP, but the uploader doesn’t support OHTTP.

pub fn http_status(code: i32) -> UploadResult

A HTTP response code.

This can still indicate an error, depending on the status code.

pub fn done() -> UploadResult

This uploader is done.