Trait glean::net::PingUploader

pub trait PingUploader:
    Debug
    + Send
    + Sync {
    // Required method
    fn upload(&self, upload_request: CapablePingUploadRequest) -> UploadResult;
}

A description of a component used to upload pings.

Required Methods

fn upload(&self, upload_request: CapablePingUploadRequest) -> UploadResult

Uploads a ping to a server.

Arguments

url - the URL path to upload the data to.

body - the serialized text data to send.

headers - a vector of tuples containing the headers to send with the request, i.e. (Name, Value).