Module glean::private

The different metric types supported by the Glean SDK to handle data.

Structs

BooleanMetric

A boolean metric.

CounterMetric

A counter metric.

CustomDistributionMetric

A custom distribution metric.

Datetime

Representation of a date, time and timezone.

DatetimeMetric

A datetime metric.

DenominatorMetric

A Denominator metric (a kind of count shared among Rate metrics).

DualLabeledCounterMetric

A dual labled metric

EventMetric

Developer-facing API for recording event metrics.

LabeledMetric

A labeled metric.

LocalCustomDistribution

Experimental: A histogram buffer associated with a specific instance of a CustomDistributionMetric.

LocalMemoryDistribution

Experimental: A histogram buffer associated with a specific instance of a MemoryDistributionMetric.

LocalTimingDistribution

Experimental: A histogram buffer associated with a specific instance of a TimingDistributionMetric.

MemoryDistributionMetric

A memory distribution metric.

NumeratorMetric

Developer-facing API for recording rate metrics with external denominators.

ObjectMetric

Developer-facing API for recording object metrics.

PingType

A ping is a bundle of related metrics, gathered in a payload to be transmitted.

QuantityMetric

A quantity metric.

RateMetric

A rate metric.

RecordedExperiment

Deserialized experiment data.

StringListMetric

A string list metric.

StringMetric

A string metric.

TextMetric

A text metric.

TimespanMetric

A timespan metric.

TimingDistributionMetric

A timing distribution metric.

UrlMetric

A URL metric.

UuidMetric

An UUID metric.

Traits

AllowLabeled

Trait for metrics that can be nested inside a labeled metric.