Struct glean::private::CustomDistributionMetric

pub struct CustomDistributionMetric { /* private fields */ }

A custom distribution metric.

Implementations

impl CustomDistributionMetric

pub fn new( meta: CommonMetricData, range_min: i64, range_max: i64, bucket_count: i64, histogram_type: HistogramType, ) -> CustomDistributionMetric

Creates a new memory distribution metric.

pub fn accumulate_samples(&self, samples: Vec<i64>)

Accumulates the provided signed samples in the metric.

This is required so that the platform-specific code can provide us with 64 bit signed integers if no u64 comparable type is available. This will take care of filtering and reporting errors for any provided negative sample.

Arguments

samples - The vector holding the samples to be recorded by the metric.

Notes

Discards any negative value in samples and report an ErrorType::InvalidValue for each of them.

pub fn accumulate_single_sample(&self, sample: i64)

Accumulates precisely one signed sample and appends it to the metric.

Signed is required so that the platform-specific code can provide us with a 64 bit signed integer if no u64 comparable type is available. This will take care of filtering and reporting errors.

Arguments

sample - The singular sample to be recorded by the metric.

Notes

Discards any negative value of sample and reports an ErrorType::InvalidValue.

pub fn test_get_num_recorded_errors(&self, error: ErrorType) -> i32

Exported for test purposes.

Gets the number of recorded errors for the given metric and error type.

Arguments

error - The type of error

Returns

The number of errors reported.

pub fn start_buffer(&self) -> LocalCustomDistribution<'_>

Experimental: Start a new histogram buffer associated with this custom distribution metric.

A histogram buffer accumulates in-memory. Data is recorded into the metric on drop.