Struct glean::private::DenominatorMetric

pub struct DenominatorMetric { /* private fields */ }

A Denominator metric (a kind of count shared among Rate metrics).

Used to count things. The value can only be incremented, not decremented.

Implementations

impl DenominatorMetric

pub fn new( meta: CommonMetricData, numerators: Vec<CommonMetricData>, ) -> DenominatorMetric

Creates a new denominator metric.

pub fn add(&self, amount: i32)

Increases the denominator by amount.

Arguments

glean - The Glean instance this metric belongs to.

amount - The amount to increase by. Should be positive.

Notes

Logs an error if the amount is 0 or negative.

pub fn test_get_value(&self, ping_name: Option<String>) -> Option<i32>

Test-only API (exported for FFI purposes).

Gets the currently stored value as an integer.

This doesn’t clear the stored value.

Arguments

ping_name - the optional name of the ping to retrieve the metric for. Defaults to the first value in send_in_pings.

Returns

The stored value or None if nothing stored.

pub fn test_get_num_recorded_errors(&self, error: ErrorType) -> i32

Exported for test purposes.

Gets the number of recorded errors for the given metric and error type.

Arguments

error - The type of error

Returns

The number of errors reported.