Struct glean::private::TimespanMetric

pub struct TimespanMetric(_);

Developer-facing API for recording timespan metrics.

Instances of this class type are automatically generated by the parsers at build time, allowing developers to record values that were previously registered in the metrics.yaml file.

Implementations

impl TimespanMetric

pub fn new(meta: CommonMetricData, time_unit: TimeUnit) -> Self

The public constructor used by automatically generated metrics.

impl TimespanMetric

pub fn start(&self)

See glean_core::traits::Timespan::start

pub fn stop(&self)

See glean_core::traits::Timespan::stop

pub fn cancel(&self)

See glean_core::traits::Timespan::cancel

pub fn set_raw(&self, __arg1: Duration)

See glean_core::traits::Timespan::set_raw

pub fn test_get_value<'a, S: Into<Option<&'a str>>>(
&self,
__arg1: S
) -> Option<u64>

See glean_core::traits::Timespan::test_get_value

pub fn test_get_num_recorded_errors<'a, S: Into<Option<&'a str>>>(
&self,
__arg1: ErrorType,
__arg2: S
) -> i32

See glean_core::traits::Timespan::test_get_num_recorded_errors