Struct glean::ConfigurationBuilder

pub struct ConfigurationBuilder {

    pub upload_enabled: bool,
    pub data_path: PathBuf,
    pub application_id: String,
    pub max_events: Option<usize>,
    pub delay_ping_lifetime_io: bool,
    pub server_endpoint: Option<String>,
    pub uploader: Option<Box<dyn PingUploader + 'static>>,
    pub use_core_mps: bool,
    pub trim_data_to_registered_pings: bool,
    pub log_level: Option<LevelFilter>,
    pub rate_limit: Option<PingRateLimit>,
    pub enable_event_timestamps: bool,
    pub experimentation_id: Option<String>,
    pub enable_internal_pings: bool,
    pub ping_schedule: HashMap<String, Vec<String>>,
    pub ping_lifetime_threshold: usize,
    pub ping_lifetime_max_time: Duration,

}

Configuration builder.

Let’s you build a configuration from the required fields and let you set optional fields individually.

Fields
upload_enabled: bool

Required: Whether upload should be enabled.

data_path: PathBuf

Required: Path to a directory to store all data in.

application_id: String

Required: The application ID (will be sanitized during initialization).

max_events: Option<usize>

Optional: The maximum number of events to store before sending a ping containing events. Default: None

delay_ping_lifetime_io: bool

Optional: Whether Glean should delay persistence of data from metrics with ping lifetime. Default: false

server_endpoint: Option<String>

Optional: The server pings are sent to. Default: None

uploader: Option<Box<dyn PingUploader + 'static>>

Optional: The instance of the uploader used to send pings. Default: None

use_core_mps: bool

Optional: Whether Glean should schedule “metrics” pings for you. Default: false

trim_data_to_registered_pings: bool

Optional: Whether Glean should limit its storage to only that of registered pings. Unless you know that all your and your libraries’ pings are appropriately registered before init, you shouldn’t use this. Default: false

log_level: Option<LevelFilter>

Optional: The internal logging level. Default: None

rate_limit: Option<PingRateLimit>

Optional: The internal ping upload rate limit. Default: None

enable_event_timestamps: bool

Whether to add a wallclock timestamp to all events.

experimentation_id: Option<String>

An experimentation identifier derived by the application to be sent with all pings, it should be noted that this has an underlying StringMetric and so should conform to the limitations that StringMetric places on length, etc.

enable_internal_pings: bool

Whether to enable internal pings. Default: true

ping_schedule: HashMap<String, Vec<String>>

A ping schedule map. Maps a ping name to a list of pings to schedule along with it. Only used if the ping’s own ping schedule list is empty.

ping_lifetime_threshold: usize

Write count threshold when to auto-flush. 0 disables it.

ping_lifetime_max_time: Duration

After what time to auto-flush. 0 disables it.

Implementations

impl Builder

pub fn new<P: Into<PathBuf>, S: Into<String>>( upload_enabled: bool, data_path: P, application_id: S, ) -> Self

A new configuration builder.

pub fn build(self) -> Configuration

Generate the full configuration.

pub fn with_max_events(self, max_events: usize) -> Self

Set the maximum number of events to store before sending a ping containing events.

pub fn with_delay_ping_lifetime_io(self, value: bool) -> Self

Set whether Glean should delay persistence of data from metrics with ping lifetime.

pub fn with_server_endpoint<S: Into<String>>(self, server_endpoint: S) -> Self

Set the server pings are sent to.

pub fn with_uploader<U: PingUploader + 'static>(self, uploader: U) -> Self

Set the instance of the uploader used to send pings.

pub fn with_use_core_mps(self, value: bool) -> Self

Set whether Glean should schedule “metrics” pings for you.

pub fn with_trim_data_to_registered_pings(self, value: bool) -> Self

Set whether Glean should limit its storage to only that of registered pings.

pub fn with_event_timestamps(self, value: bool) -> Self

Set whether to add a wallclock timestamp to all events (experimental).

pub fn with_experimentation_id(self, value: String) -> Self

Set whether to add a wallclock timestamp to all events (experimental).

pub fn with_internal_pings(self, value: bool) -> Self

Set whether to enable internal pings.

pub fn with_ping_schedule(self, value: HashMap<String, Vec<String>>) -> Self

Set the ping schedule map.

pub fn with_ping_lifetime_threshold(self, value: usize) -> Self

Write count threshold when to auto-flush. 0 disables it.

pub fn with_ping_lifetime_max_time(self, value: Duration) -> Self

After what time to auto-flush. 0 disables it.