Module glean::traits

API definitions for the different metric types supported by the Glean SDK.

Individual metric types implement this trait to expose the specific metrics API. It can be used by wrapping implementations to guarantee API conformance.

Modules

__serde

Serde

Structs

ObjectError

This type represents all possible errors that can occur when serializing or deserializing an object from/to JSON.

Enums

EventRecordingError

The possible errors when parsing to an extra key.

HistogramType

Different kinds of histograms.

NoExtraKeys

Default of no extra keys for events.

Traits

Boolean

A description for the BooleanMetric type.

Counter

A description for the CounterMetric type.

CustomDistribution

A description for the CustomDistributionMetric type.

Datetime

A description for the DatetimeMetric type.

DualLabeledCounter

A description for the DualLabeledCounterMetric type.

Event

A description for the EventMetric type.

ExtraKeys

Extra keys for events.

Labeled

A description for the LabeledMetric type.

MemoryDistribution

A description for the MemoryDistributionMetric type.

Numerator

A description for the NumeratorMetric subtype of the RateMetric type.

ObjectSerialize

An object that can be serialized into JSON.

Ping

A description for the PingType type.

Quantity

A description for the QuantityMetric type.

Rate

A description for the RateMetric type.

String

A description for the StringMetric type.

StringList

A description for the StringListMetric type.

Text

A description for the TextMetric type.

Timespan

A description for the TimespanMetric type.

TimingDistribution

A description for the TimingDistributionMetric type.

Url

A description for the UrlMetric type.

Uuid

A description for the UuidMetric type.