Trait glean::traits::CustomDistribution

pub trait CustomDistribution: TestGetValue<Output = DistributionData> {
    // Required methods
    fn accumulate_samples_signed(&self, samples: Vec<i64>);
    fn accumulate_single_sample_signed(&self, sample: i64);
    fn test_get_num_recorded_errors(&self, error: ErrorType) -> i32;
}

A description for the CustomDistributionMetric type.

When changing this trait, make sure all the operations are implemented in the related type in ../metrics/.

Required Methods

fn accumulate_samples_signed(&self, samples: Vec<i64>)

Accumulates the provided signed samples in the metric.

This is required so that the platform-specific code can provide us with 64 bit signed integers if no u64 comparable type is available. This will take care of filtering and reporting errors for any provided negative sample.

Arguments

samples - The vector holding the samples to be recorded by the metric.

Notes

Discards any negative value in samples and report an ErrorType::InvalidValue for each of them.

fn accumulate_single_sample_signed(&self, sample: i64)

Accumulates precisely one signed sample in the metric.

This is required so that the platform-specific code can provide us with a 64 bit signed integer if no u64 comparable type is available. This will take care of filtering and reporting errors.

Arguments

sample - The singular sample to be recorded by the metric.

Notes

Discards any negative value of sample and reports an ErrorType::InvalidValue.

fn test_get_num_recorded_errors(&self, error: ErrorType) -> i32

Exported for test purposes.

Gets the number of recorded errors for the given error type.

Arguments

error - The type of error

Returns

The number of errors recorded.