Trait glean::traits::Event

pub trait Event: TestGetValue<Output = Vec<RecordedEvent>> {
    type Extra: ExtraKeys;

    // Required method
    fn record<M>(&self, extra: M)
       where M: Into<Option<Self::Extra>>;
}

A description for the EventMetric type.

When changing this trait, make sure all the operations are implemented in the related type in ../metrics/.

Required Associated Types

type Extra: ExtraKeys

The type of the allowed extra keys for this event.

Required Methods

fn record<M>(&self, extra: M)
where M: Into<Option<Self::Extra>>,

Records an event.

Arguments

extra - (optional) An object for the extra keys.

Object Safety

This trait is not object safe.