Trait glean::traits::ExtraKeys

pub trait ExtraKeys {
    const ALLOWED_KEYS: &'static [&'static str];

    fn into_ffi_extra(self) -> HashMap<i32, String, RandomState>;
}

Extra keys for events.

Extra keys need to be pre-defined and map to a string representation.

For user-defined EventMetrics these will be defined as structs. Each extra key will be a field in that struct. Each field will correspond to an entry in the ALLOWED_KEYS list. The Glean SDK requires the keys as strings for submission in pings, whereas in code we want to provide users a type to work with (e.g. to avoid typos or misuse of the API).

Associated Constants

const ALLOWED_KEYS: &'static [&'static str]

List of allowed extra keys as strings.

Required methods

fn into_ffi_extra(self) -> HashMap<i32, String, RandomState>

Convert the event extras into 2 lists:

The list of extra key indices. Unset keys will be skipped.

The list of extra values.