Trait glean::traits::Jwe

pub trait Jwe {
    fn set_with_compact_representation<S>(&self, value: S)
    where
        S: Into<String>;

    fn set<S>(
        &self, 
        header: S, 
        key: S, 
        init_vector: S, 
        cipher_text: S, 
        auth_tag: S
    )
    where
        S: Into<String>;

    fn test_get_value<'a, S>(&self, ping_name: S) -> Option<String>
    where
        S: Into<Option<&'a str>>;

    fn test_get_value_as_json_string<'a, S>(
        &self, 
        ping_name: S
    ) -> Option<String>
    where
        S: Into<Option<&'a str>>;
}

A description for the JweMetric type.

When changing this trait, make sure all the operations are implemented in the related type in ../metrics/.

Required methods

fn set_with_compact_representation<S>(&self, value: S) where
S: Into<String>,

Sets to the specified JWE value.

Arguments

value - the compact representation of a JWE value.

fn set<S>(&self, header: S, key: S, init_vector: S, cipher_text: S, auth_tag: S) where
S: Into<String>,

Builds a JWE value from its elements and set to it.

Arguments

header - the JWE Protected Header element.

key - the JWE Encrypted Key element.

init_vector - the JWE Initialization Vector element.

cipher_text - the JWE Ciphertext element.

auth_tag - the JWE Authentication Tag element.

fn test_get_value<'a, S>(&self, ping_name: S) -> Option<String> where
S: Into<Option<&'a str>>,

Exported for test purposes.

Gets the currently stored value as a string.

This doesn’t clear the stored value.

Arguments

ping_name - represents the optional name of the ping to retrieve the metric for. Defaults to the first value in send_in_pings.

fn test_get_value_as_json_string<'a, S>(&self, ping_name: S) -> Option<String> where
S: Into<Option<&'a str>>,

Exported for test purposes.

Gets the currently stored JWE as a JSON String of the serialized value.

This doesn’t clear the stored value.

Arguments

ping_name - represents the optional name of the ping to retrieve the metric for. Defaults to the first value in send_in_pings.