Trait glean::traits::StringList

pub trait StringList: TestGetValue<Output = Vec<String>> {
    // Required methods
    fn add<S>(&self, value: S)
       where S: Into<String>;
    fn set(&self, value: Vec<String>);
    fn test_get_num_recorded_errors(&self, error: ErrorType) -> i32;
}

A description for the StringListMetric type.

When changing this trait, make sure all the operations are implemented in the related type in ../metrics/.

Required Methods

fn add<S>(&self, value: S)
where S: Into<String>,

Adds a new string to the list.

Arguments

value - The string to add.

Notes

Truncates the value if it is longer than MAX_STRING_LENGTH bytes and logs an error.

fn set(&self, value: Vec<String>)

Sets to a specific list of strings.

Arguments

value - The list of string to set the metric to.

Notes

If passed an empty list, records an error and returns. Truncates the list if it is longer than MAX_LIST_LENGTH and logs an error. Truncates any value in the list if it is longer than MAX_STRING_LENGTH and logs an error.

fn test_get_num_recorded_errors(&self, error: ErrorType) -> i32

Exported for test purposes.

Gets the number of recorded errors for the given error type.

Arguments

error - The type of error

Returns

The number of errors recorded.

Object Safety

This trait is not object safe.