Trait glean::traits::Timespan

pub trait Timespan {
    fn start(&self);

    fn stop(&self);

    fn cancel(&self);

    fn test_get_value<'a, S>(&self, ping_name: S) -> Option<u64>
    where
        S: Into<Option<&'a str>>;

    fn test_get_num_recorded_errors<'a, S>(
        &self, 
        error: ErrorType, 
        ping_name: S
    ) -> i32
    where
        S: Into<Option<&'a str>>;
}

A description for the TimespanMetric type.

When changing this trait, make sure all the operations are implemented in the related type in ../metrics/.

Required methods

fn start(&self)

Starts tracking time for the provided metric.

This uses an internal monotonic timer.

This records an error if it's already tracking time (i.e. start was already called with no corresponding stop): in that case the original start time will be preserved.

fn stop(&self)

Stops tracking time for the provided metric. Sets the metric to the elapsed time.

This will record an error if no start was called.

fn cancel(&self)

Aborts a previous start call. No error is recorded if no start was called.

fn test_get_value<'a, S>(&self, ping_name: S) -> Option<u64> where
S: Into<Option<&'a str>>,

Exported for test purposes.

Gets the currently stored value as an integer.

This doesn't clear the stored value.

Arguments

ping_name - represents the optional name of the ping to retrieve the metric for. Defaults to the first value in send_in_pings.

fn test_get_num_recorded_errors<'a, S>(
&self,
error: ErrorType,
ping_name: S
) -> i32 where
S: Into<Option<&'a str>>,

Exported for test purposes.

Gets the number of recorded errors for the given metric and error type.

Arguments

error - The type of error

ping_name - represents the optional name of the ping to retrieve the metric for. Defaults to the first value in send_in_pings.

Returns

The number of errors reported.

Loading content...