Trait glean::traits::TimingDistribution

pub trait TimingDistribution {
    fn start(&self) -> u64;

    fn stop_and_accumulate(&self, id: u64);

    fn cancel(&self, id: u64);

    fn test_get_value<'a, S>(&self, ping_name: S) -> Option<DistributionData>
    where
        S: Into<Option<&'a str>>;

    fn test_get_num_recorded_errors<'a, S>(
        &self, 
        error: ErrorType, 
        ping_name: S
    ) -> i32
    where
        S: Into<Option<&'a str>>;
}

A description for the TimingDistributionMetric type.

When changing this trait, make sure all the operations are implemented in the related type in ../metrics/.

Required methods

fn start(&self) -> u64

Start tracking time for the provided metric. Multiple timers can run simultaneously.

Returns

A unique TimerId for the new timer.

fn stop_and_accumulate(&self, id: u64)

Stops tracking time for the provided metric and associated timer id.

Adds a count to the corresponding bucket in the timing distribution. This will record an error if no start was called.

Arguments

id - The TimerId to associate with this timing. This allows for concurrent timing of events associated with different ids to the same timespan metric.

fn cancel(&self, id: u64)

Aborts a previous start call. No error is recorded if no start was called.

Arguments

id - The TimerId to associate with this timing. This allows for concurrent timing of events associated with different ids to the same timing distribution metric.

fn test_get_value<'a, S>(&self, ping_name: S) -> Option<DistributionData> where
S: Into<Option<&'a str>>,

Exported for test purposes.

Gets the currently stored value of the metric.

This doesn’t clear the stored value.

Arguments

ping_name - represents the optional name of the ping to retrieve the metric for. Defaults to the first value in send_in_pings.

fn test_get_num_recorded_errors<'a, S>(
&self,
error: ErrorType,
ping_name: S
) -> i32 where
S: Into<Option<&'a str>>,

Exported for test purposes.

Gets the number of recorded errors for the given error type.

Arguments

error - The type of error

ping_name - represents the optional name of the ping to retrieve the metric for. Defaults to the first value in send_in_pings.

Returns

The number of errors recorded.