Module glean::private

The different metric types supported by the Glean SDK to handle data.

Structs

BooleanMetric
Developer-facing API for recording boolean metrics.

CounterMetric
Developer-facing API for recording counter metrics.

CustomDistributionMetric
Developer-facing API for recording custom distribution metrics.

DatetimeMetric
Developer-facing API for recording datetime metrics.

DenominatorMetric
Developer-facing API for recording counter metrics that are acting as external denominators for rate metrics.

EventMetric
Developer-facing API for recording event metrics.

LabeledMetric
Developer-facing API for recording labeled metrics.

MemoryDistributionMetric
Developer-facing API for recording memory distribution metrics.

NumeratorMetric
Developer-facing API for recording rate metrics with external denominators.

PingType
A ping is a bundle of related metrics, gathered in a payload to be transmitted.

QuantityMetric
Developer-facing API for recording quantity metrics.

RateMetric
Developer-facing API for recording rate metrics.

RecordedExperimentData
Deserialized experiment data.

StringListMetric
Developer-facing API for recording string list metrics.

StringMetric
Developer-facing API for recording string metrics.

TimespanMetric
Developer-facing API for recording timespan metrics.

TimingDistributionMetric
Developer-facing API for recording timing distribution metrics.

UrlMetric
Developer-facing API for recording URL metrics.

UuidMetric
Developer-facing API for recording UUID metrics.

Traits

AllowLabeled
Marker trait for metrics that can be nested inside a labeled metric.

Type Definitions

Datetime
A datetime type.