Trait glean::TestGetValue

pub trait TestGetValue {
    type Output;

    // Required method
    fn test_get_value(&self, ping_name: Option<String>) -> Option<Self::Output>;
}

TestGetValue describes an interface for retrieving the value for a given metric

Required Associated Types

type Output

The output type of test_get_value

Required Methods

fn test_get_value(&self, ping_name: Option<String>) -> Option<Self::Output>

Test-only API (exported for FFI purposes).

Returns the currently stored value of the appropriate type for the given metric.

This doesn’t clear the stored value.

Arguments

ping_name - the optional name of the ping to retrieve the metric for. Defaults to the first value in send_in_pings.

Returns

The stored value or None if nothing stored.