Module glean::traits

API definitions for the different metric types supported by the Glean SDK.

Individual metric types implement this trait to expose the specific metrics API. It can be used by wrapping implementations to guarantee API conformance.

Enums

EventRecordingError
The possible errors when parsing to an extra key.

HistogramType
Different kinds of histograms.

NoExtraKeys
Default of no extra keys for events.

Traits

Boolean
A description for the BooleanMetric type.

Counter
A description for the CounterMetric type.

CustomDistribution
A description for the CustomDistributionMetric type.

Datetime
A description for the DatetimeMetric type.

Event
A description for the EventMetric type.

ExtraKeys
Extra keys for events.

Jwe
A description for the JweMetric type.

Labeled
A description for the LabeledMetric type.

MemoryDistribution
A description for the MemoryDistributionMetric type.

Ping
A description for the PingType type.

Quantity
A description for the QuantityMetric type.

String
A description for the StringMetric type.

StringList
A description for the StringListMetric type.

Timespan
A description for the TimespanMetric type.

TimingDistribution
A description for the TimingDistributionMetric type.

Uuid
A description for the UuidMetric type.