Trait glean::traits::Event

pub trait Event {
    type Extra: ExtraKeys;
    fn record<M>(&self, extra: M)
    where
        M: Into<Option<HashMap<Self::Extra, String, RandomState>>>;

    fn test_get_value<'a, S>(&self, ping_name: S) -> Option<Vec<RecordedEvent>>
    where
        S: Into<Option<&'a str>>;

    fn test_get_num_recorded_errors<'a, S>(
        &self, 
        error: ErrorType, 
        ping_name: S
    ) -> i32
    where
        S: Into<Option<&'a str>>;
}

A description for the EventMetric type.

When changing this trait, make sure all the operations are implemented in the related type in ../metrics/.

Associated Types

type Extra: ExtraKeys

The type of the allowed extra keys for this event.

Loading content...
Required methods

fn record<M>(&self, extra: M) where
M: Into<Option<HashMap<Self::Extra, String, RandomState>>>,

Records an event.

Arguments

extra - A HashMap of (key, value) pairs. The key is one of the allowed extra keys as set in the metric defintion. If a wrong key is used or a value is larger than allowed, an error is reported and no event is recorded.

fn test_get_value<'a, S>(&self, ping_name: S) -> Option<Vec<RecordedEvent>> where
S: Into<Option<&'a str>>,

Exported for test purposes.

Get the vector of currently stored events for this event metric.

This doesn't clear the stored value.

Arguments

ping_name - represents the optional name of the ping to retrieve the metric for. Defaults to the first value in send_in_pings.

fn test_get_num_recorded_errors<'a, S>(
&self,
error: ErrorType,
ping_name: S
) -> i32 where
S: Into<Option<&'a str>>,

Exported for test purposes.

Gets the number of recorded errors for the given metric and error type.

Arguments

error - The type of error

ping_name - represents the optional name of the ping to retrieve the metric for. Defaults to the first value in send_in_pings.

Returns

The number of errors reported.

Loading content...