Trait glean::traits::Numerator

pub trait Numerator {
    fn add_to_numerator(&self, amount: i32);

    fn test_get_value<'a, S>(&self, ping_name: S) -> Option<(i32, i32)>
    where
        S: Into<Option<&'a str>>;

    fn test_get_num_recorded_errors<'a, S>(
        &self, 
        error: ErrorType, 
        ping_name: S
    ) -> i32
    where
        S: Into<Option<&'a str>>;
}

A description for the NumeratorMetric subtype of the RateMetric type.

Required methods

fn add_to_numerator(&self, amount: i32)

Increases the numerator by amount.

Arguments

amount - The amount to increase by. Should be non-negative.

Notes

Logs an error if the amount is negative.

fn test_get_value<'a, S>(&self, ping_name: S) -> Option<(i32, i32)> where
S: Into<Option<&'a str>>,

Exported for test purposes.

Gets the currently stored value as a pair of integers.

Arguments

ping_name - the optional name of the ping to retrieve the metric for. Defaults to the first value in send_in_pings.

This doesn’t clear the stored value.

fn test_get_num_recorded_errors<'a, S>(
&self,
error: ErrorType,
ping_name: S
) -> i32 where
S: Into<Option<&'a str>>,

Exported for test purposes.

Gets the number of recorded errors for the given metric and error type.

Arguments

error - The type of error

ping_name - the optional name of the ping to retrieve the metric for. Defaults to the first value in send_in_pings.

Returns

The number of errors reported.