Trait glean::traits::StringList

pub trait StringList {
    fn add<S>(&self, value: S)
    where
        S: Into<String>;

    fn set(&self, value: Vec<String>);

    fn test_get_value<'a, S>(&self, ping_name: S) -> Option<Vec<String>>
    where
        S: Into<Option<&'a str>>;

    fn test_get_num_recorded_errors<'a, S>(
        &self, 
        error: ErrorType, 
        ping_name: S
    ) -> i32
    where
        S: Into<Option<&'a str>>;
}

A description for the StringListMetric type.

When changing this trait, make sure all the operations are implemented in the related type in ../metrics/.

Required methods

fn add<S>(&self, value: S) where
S: Into<String>,

Adds a new string to the list.

Arguments

value - The string to add.

Notes

Truncates the value if it is longer than MAX_STRING_LENGTH bytes and logs an error.

fn set(&self, value: Vec<String>)

Sets to a specific list of strings.

Arguments

value - The list of string to set the metric to.

Notes

If passed an empty list, records an error and returns. Truncates the list if it is longer than MAX_LIST_LENGTH and logs an error. Truncates any value in the list if it is longer than MAX_STRING_LENGTH and logs an error.

fn test_get_value<'a, S>(&self, ping_name: S) -> Option<Vec<String>> where
S: Into<Option<&'a str>>,

Exported for test purposes.

Gets the currently-stored values.

This doesn't clear the stored value.

Arguments

ping_name - represents the optional name of the ping to retrieve the metric for. Defaults to the first value in send_in_pings.

fn test_get_num_recorded_errors<'a, S>(
&self,
error: ErrorType,
ping_name: S
) -> i32 where
S: Into<Option<&'a str>>,

Exported for test purposes.

Gets the number of recorded errors for the given error type.

Arguments

error - The type of error

ping_name - represents the optional name of the ping to retrieve the metric for. Defaults to the first value in send_in_pings.

Returns

The number of errors recorded.

Loading content...